refactor(neon_framework): move storage keys into the storage library
