+++ packages/neon/neon_news/lib/src/options.dart
@@ -1,6 +1,7 @@
 import 'package:neon_framework/platform.dart';
 import 'package:neon_framework/settings.dart';
 import 'package:neon_framework/sort_box.dart';
+import 'package:neon_framework/storage.dart';
 import 'package:neon_news/l10n/localizations.dart';
 import 'package:neon_news/src/blocs/articles.dart';
