+++ packages/neon_framework/lib/settings.dart
@@ -2,5 +2,4 @@ export 'package:neon_framework/src/models/label_builder.dart';
 export 'package:neon_framework/src/settings/models/option.dart';
 export 'package:neon_framework/src/settings/models/options_category.dart';
 export 'package:neon_framework/src/settings/models/options_collection.dart';
-export 'package:neon_framework/src/settings/models/storage.dart' show Storable;
 export 'package:neon_framework/src/settings/widgets/settings_list.dart';
