+++ packages/neon_framework/lib/src/storage/settings_store.dart
@@ -1,5 +1,5 @@
 import 'package:meta/meta.dart';
-import 'package:neon_framework/src/settings/models/storage.dart';
+import 'package:neon_framework/src/storage/keys.dart';
 import 'package:neon_framework/src/storage/storage_manager.dart';
 
 /// Storage interface used by `Option`s.
