+++ packages/neon_framework/lib/src/storage/single_value_store.dart
@@ -1,6 +1,6 @@
 import 'package:meta/meta.dart';
-import 'package:neon_framework/src/settings/models/storage.dart';
-import 'package:neon_framework/storage.dart';
+import 'package:neon_framework/src/storage/keys.dart';
+import 'package:neon_framework/src/storage/storage_manager.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 /// A storage that itself is a single entry of a key value store.
