+++ packages/neon_framework/lib/storage.dart
@@ -3,6 +3,7 @@
 /// The `NeonStorage` manages all storage backends.
 library;
 
+export 'package:neon_framework/src/storage/keys.dart' show Storable;
 export 'package:neon_framework/src/storage/request_cache.dart';
 export 'package:neon_framework/src/storage/settings_store.dart' show SettingsStorage;
 export 'package:neon_framework/src/storage/single_value_store.dart' show KeyValueStorage;
