+++ packages/neon_framework/test/options_collection_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:neon_framework/settings.dart';
-import 'package:neon_framework/src/settings/models/storage.dart';
+import 'package:neon_framework/src/storage/keys.dart';
 import 'package:neon_framework/src/storage/settings_store.dart';
 import 'package:neon_framework/testing.dart';
