+++ packages/neon_framework/test/settings_test.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:neon_framework/src/settings/models/option.dart';
-import 'package:neon_framework/src/settings/models/storage.dart';
 import 'package:neon_framework/src/settings/widgets/option_settings_tile.dart';
+import 'package:neon_framework/storage.dart';
 import 'package:neon_framework/testing.dart';
 
 enum StorageKey implements Storable {
